script damage on dead characters fix

+++ src/dg_misc.c
@@ -373,6 +373,11 @@ void script_damage(char_data *vict, char_data *killer, int level, int dam_type, 
 	
 	double dam;
 	
+	// no point damaging the dead
+	if (IS_DEAD(vict)) {
+		return;
+	}
+	
 	if (IS_IMMORTAL(vict) && (modifier > 0)) {
 		msg_to_char(vict, "Being the cool immortal you are, you sidestep a trap, obviously placed to kill you.\r\n");
 		return;
